+++ app/src/main/java/com/appyhome/appyproduct/mvvm/data/model/api/product/DeleteWishListRequest.java
@@ -7,13 +7,13 @@ import com.google.gson.annotations.SerializedName;
 public class DeleteWishListRequest {
     @Expose
     @SerializedName("product_id")
-    public int product_id;
+    public long product_id;
 
     @Expose
     @SerializedName("variant_id")
     public int variant_id;
 
-    public DeleteWishListRequest(int idP, int idV) {
+    public DeleteWishListRequest(long idP, int idV) {
         product_id = idP;
         variant_id = idV;
     }
