+++ app/src/main/java/com/appyhome/appyproduct/mvvm/ui/appyproduct/product/list/ProductListNavigator.java
@@ -17,7 +17,7 @@ public interface ProductListNavigator {
 
     void showProducts(OrderedRealmCollection<Product> result);
 
-    void getAllFavorites_Done(ArrayList<Integer> listId);
+    void getAllFavorites_Done(ArrayList<Long> listId);
 
     void showEmptyProducts();
